changed '/' position , after port number, in order to correctly show suggestions for addresses containing port numbers
