+++ samples/browser/src/main/java/org/mozilla/samples/browser/BrowserActivity.kt
@@ -10,6 +10,7 @@ import android.os.Bundle
 import android.util.AttributeSet
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
 import mozilla.components.browser.tabstray.BrowserTabsTray
 import mozilla.components.concept.engine.EngineView
 import mozilla.components.concept.tabstray.TabsTray
@@ -19,6 +20,10 @@ import mozilla.components.support.utils.SafeIntent
 import org.mozilla.samples.browser.ext.components
 
 open class BrowserActivity : AppCompatActivity(), ComponentCallbacks2 {
+
+    open fun createBrowserFragment(sessionId: String?): Fragment =
+        BrowserFragment.create(sessionId)
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
@@ -26,7 +31,7 @@ open class BrowserActivity : AppCompatActivity(), ComponentCallbacks2 {
         if (savedInstanceState == null) {
             val sessionId = SafeIntent(intent).getStringExtra(IntentProcessor.ACTIVE_SESSION_ID)
             supportFragmentManager?.beginTransaction()?.apply {
-                replace(R.id.container, BrowserFragment.create(sessionId))
+                replace(R.id.container, createBrowserFragment(sessionId))
                 commit()
             }
         }
