+++ samples/browser/src/main/java/org/mozilla/samples/browser/BrowserFragment.kt
@@ -4,105 +4,28 @@
 
 package org.mozilla.samples.browser
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import kotlinx.android.synthetic.main.fragment_browser.view.*
-import mozilla.components.browser.session.SelectionAwareSessionObserver
-import mozilla.components.browser.session.Session
-import mozilla.components.feature.app.links.AppLinksFeature
 import mozilla.components.feature.awesomebar.AwesomeBarFeature
 import mozilla.components.feature.awesomebar.provider.SearchSuggestionProvider
-import mozilla.components.feature.contextmenu.ContextMenuCandidate
-import mozilla.components.feature.contextmenu.ContextMenuFeature
-import mozilla.components.feature.customtabs.CustomTabsToolbarFeature
-import mozilla.components.feature.downloads.DownloadsFeature
-import mozilla.components.feature.downloads.manager.FetchDownloadManager
-import mozilla.components.feature.prompts.PromptFeature
-import mozilla.components.feature.session.CoordinateScrollingFeature
-import mozilla.components.feature.session.SessionFeature
-import mozilla.components.feature.session.SwipeRefreshFeature
 import mozilla.components.feature.session.ThumbnailsFeature
-import mozilla.components.feature.session.WindowFeature
-import mozilla.components.feature.sitepermissions.SitePermissionsFeature
 import mozilla.components.feature.tabs.toolbar.TabsToolbarFeature
 import mozilla.components.feature.toolbar.ToolbarAutocompleteFeature
-import mozilla.components.feature.toolbar.ToolbarFeature
 import mozilla.components.support.base.feature.BackHandler
-import mozilla.components.support.base.feature.LifecycleAwareFeature
 import mozilla.components.support.base.feature.ViewBoundFeatureWrapper
-import mozilla.components.support.base.log.logger.Logger
-import mozilla.components.support.ktx.android.arch.lifecycle.addObservers
-import org.mozilla.samples.browser.downloads.DownloadService
 import org.mozilla.samples.browser.ext.components
-import org.mozilla.samples.browser.integration.FindInPageIntegration
 import org.mozilla.samples.browser.integration.ReaderViewIntegration
 
-class BrowserFragment : Fragment(), BackHandler {
-    private val sessionFeature = ViewBoundFeatureWrapper<SessionFeature>()
-    private val toolbarFeature = ViewBoundFeatureWrapper<ToolbarFeature>()
-    private val customTabsToolbarFeature = ViewBoundFeatureWrapper<CustomTabsToolbarFeature>()
-    private val downloadsFeature = ViewBoundFeatureWrapper<DownloadsFeature>()
-    private val promptFeature = ViewBoundFeatureWrapper<PromptFeature>()
-    private val findInPageIntegration = ViewBoundFeatureWrapper<FindInPageIntegration>()
-    private val sitePermissionsFeature = ViewBoundFeatureWrapper<SitePermissionsFeature>()
+class BrowserFragment : BaseBrowserFragment(), BackHandler {
     private val thumbnailsFeature = ViewBoundFeatureWrapper<ThumbnailsFeature>()
     private val readerViewFeature = ViewBoundFeatureWrapper<ReaderViewIntegration>()
-    private val swipeRefreshFeature = ViewBoundFeatureWrapper<SwipeRefreshFeature>()
-    private val appLinksFeature = ViewBoundFeatureWrapper<AppLinksFeature>()
 
     @Suppress("LongMethod")
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val layout = inflater.inflate(R.layout.fragment_browser, container, false)
-
-        layout.toolbar.setMenuBuilder(components.menuBuilder)
-
-        val sessionId = arguments?.getString(SESSION_ID)
-
-        sessionFeature.set(
-            feature = SessionFeature(
-                components.sessionManager,
-                components.sessionUseCases,
-                layout.engineView,
-                sessionId),
-            owner = this,
-            view = layout)
-
-        toolbarFeature.set(
-            feature = ToolbarFeature(
-                layout.toolbar,
-                components.sessionManager,
-                components.sessionUseCases.loadUrl,
-                components.defaultSearchUseCase,
-                sessionId),
-            owner = this,
-            view = layout)
-
-        swipeRefreshFeature.set(
-            feature = SwipeRefreshFeature(
-                components.sessionManager,
-                components.sessionUseCases.reload,
-                layout.swipeToRefresh),
-            owner = this,
-            view = layout)
-
-        val menuUpdaterFeature = object : SelectionAwareSessionObserver(components.sessionManager),
-                LifecycleAwareFeature {
-            override fun onLoadingStateChanged(session: Session, loading: Boolean) {
-                layout.toolbar.invalidateActions()
-            }
-
-            override fun onNavigationStateChanged(session: Session, canGoBack: Boolean, canGoForward: Boolean) {
-                layout.toolbar.invalidateActions()
-            }
-
-            override fun start() {
-                observeIdOrSelected(sessionId)
-            }
-        }
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+        val layout = super.onCreateView(inflater, container, savedInstanceState)
 
         ToolbarAutocompleteFeature(layout.toolbar).apply {
             addHistoryStorageProvider(components.historyStorage)
@@ -123,77 +46,6 @@ class BrowserFragment : Fragment(), BackHandler {
                 mode = SearchSuggestionProvider.Mode.MULTIPLE_SUGGESTIONS)
             .addClipboardProvider(requireContext(), components.sessionUseCases.loadUrl)
 
-        downloadsFeature.set(
-            feature = DownloadsFeature(
-                requireContext().applicationContext,
-                sessionManager = components.sessionManager,
-                fragmentManager = childFragmentManager,
-                onDownloadCompleted = { download, id ->
-                    Logger.debug("Download done. ID#$id $download")
-                },
-                downloadManager = FetchDownloadManager(
-                    requireContext().applicationContext,
-                    DownloadService::class
-                ),
-                onNeedToRequestPermissions = { permissions ->
-                    requestPermissions(permissions, REQUEST_CODE_DOWNLOAD_PERMISSIONS)
-                }),
-            owner = this,
-            view = layout)
-
-        val scrollFeature = CoordinateScrollingFeature(components.sessionManager, layout.engineView, layout.toolbar)
-
-        val contextMenuFeature = ContextMenuFeature(
-            requireFragmentManager(),
-            components.sessionManager,
-            ContextMenuCandidate.defaultCandidates(
-                requireContext(),
-                components.tabsUseCases,
-                layout),
-            layout.engineView)
-
-        promptFeature.set(
-            feature = PromptFeature(
-                fragment = this,
-                sessionManager = components.sessionManager,
-                sessionId = sessionId,
-                fragmentManager = requireFragmentManager(),
-                onNeedToRequestPermissions = { permissions ->
-                    requestPermissions(permissions, REQUEST_CODE_PROMPT_PERMISSIONS)
-                }),
-            owner = this,
-            view = layout)
-
-        val windowFeature = WindowFeature(components.engine, components.sessionManager)
-
-        sitePermissionsFeature.set(
-            feature = SitePermissionsFeature(
-                context = requireContext(),
-                sessionManager = components.sessionManager,
-                sessionId = sessionId,
-                fragmentManager = requireFragmentManager()
-            ) { permissions ->
-                requestPermissions(permissions, REQUEST_CODE_APP_PERMISSIONS)
-            },
-            owner = this,
-            view = layout
-        )
-
-        customTabsToolbarFeature.set(
-            feature = CustomTabsToolbarFeature(
-                components.sessionManager,
-                layout.toolbar,
-                sessionId,
-                components.menuBuilder,
-                closeListener = { activity?.finish() }),
-            owner = this,
-            view = layout)
-
-        findInPageIntegration.set(
-            feature = FindInPageIntegration(components.sessionManager, layout.findInPage, layout.engineView),
-            owner = this,
-            view = layout)
-
         readerViewFeature.set(
             feature = ReaderViewIntegration(
                 requireContext(),
@@ -207,27 +59,8 @@ class BrowserFragment : Fragment(), BackHandler {
             view = layout
         )
 
-        // Observe the lifecycle for supported features
-        lifecycle.addObservers(
-            scrollFeature,
-            contextMenuFeature,
-            windowFeature,
-            menuUpdaterFeature
-        )
-
         thumbnailsFeature.set(
             feature = ThumbnailsFeature(requireContext(), layout.engineView, components.sessionManager),
-            owner = this,
-            view = layout
-        )
-
-        appLinksFeature.set(
-            feature = AppLinksFeature(
-                context = requireContext(),
-                sessionManager = components.sessionManager,
-                sessionId = sessionId,
-                fragmentManager = requireFragmentManager()
-            ),
             owner = this,
             view = layout
         )
@@ -244,45 +77,14 @@ class BrowserFragment : Fragment(), BackHandler {
         }
     }
 
-    override fun onBackPressed(): Boolean {
-        return when {
-            readerViewFeature.onBackPressed() -> true
-            findInPageIntegration.onBackPressed() -> true
-            toolbarFeature.onBackPressed() -> true
-            sessionFeature.onBackPressed() -> true
-            customTabsToolbarFeature.onBackPressed() -> true
-            else -> false
-        }
-    }
+    override fun onBackPressed(): Boolean =
+        readerViewFeature.onBackPressed() || super.onBackPressed()
 
     companion object {
-        private const val SESSION_ID = "session_id"
-        private const val REQUEST_CODE_DOWNLOAD_PERMISSIONS = 1
-        private const val REQUEST_CODE_PROMPT_PERMISSIONS = 2
-        private const val REQUEST_CODE_APP_PERMISSIONS = 3
-
-        fun create(sessionId: String? = null): BrowserFragment = BrowserFragment().apply {
+        fun create(sessionId: String? = null) = BrowserFragment().apply {
             arguments = Bundle().apply {
-                putString(SESSION_ID, sessionId)
+                putSessionId(sessionId)
             }
         }
-    }
-
-    override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<String>, grantResults: IntArray) {
-        when (requestCode) {
-            REQUEST_CODE_DOWNLOAD_PERMISSIONS -> downloadsFeature.withFeature {
-                it.onPermissionsResult(permissions, grantResults)
-            }
-            REQUEST_CODE_PROMPT_PERMISSIONS -> promptFeature.withFeature {
-                it.onPermissionsResult(permissions, grantResults)
-            }
-            REQUEST_CODE_APP_PERMISSIONS -> sitePermissionsFeature.withFeature {
-                it.onPermissionsResult(permissions, grantResults)
-            }
-        }
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        promptFeature.withFeature { it.onActivityResult(requestCode, resultCode, data) }
     }
 }
